reorg some renter tests
